Tools/CLI/pyproject.toml
[build-system]
requires = [ "poetry>=0.12",]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "mdk-cli"
version = "0.1.0"
description = "mdk's comprehensive command line interface"
authors = [ "Ben Flint <bflint@popcap.com>",]

[tool.black]
line-length = 120
target-version = [ "py38",]
include = "\\.pyi?$"
exclude = "(\n  /(\n      \\.eggs         # exclude a few common directories in the\n    | \\.git          # root of the project\n    | \\.hg\n    | \\.mypy_cache\n    | \\.pytest_cache\n    | \\.tox\n    | \\.venv\n    | _build\n    | buck-out\n    | build\n    | dist\n  )/\n  #| foo.py           # also separately exclude a file named foo.py in\n                     # the root of the project\n)\n"

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.1"
flake8 = "^3.7.9"
protobuf = "^3.11.3"
requests = "^2.23.0"
requests_toolbelt = "^0.9.1"
nodeenv = "^1.3.5"
toml = "^0.10.1"
loguru = "^0.5.1"
invoke = "^1.4.1"
jsbeautifier = "^1.11.0"
jsonschema = "^3.2.0"
jinja2 = "^2.11.2"
python-box = "^5.1.0"
python-slugify = "^4.0.1"
sortedcontainers = "^2.2.2"
stringcase = "^1.2.0"
boto3 = "^1.15.2"

[tool.poetry.dev-dependencies]
pytest = "6.0.1"
pytest-xdist = "^2.1.0"
pytest-cov = "^2.10.1"
pytest-bdd = "^3.4.0"

[tool.poetry.scripts]
mdk = "mdk_cli.cli_main:main"
dia = "mdk_cli.cli_main:main"

[tool.pylint.format]
max-line-length = "120"

[tool.pylint.messages_control]
disable = "C0330, C0326"

[tool.poetry.dev-dependencies.black]
version = "^20.8b1"
allow-prereleases = true
